# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "McpAuthPayload": {
        "additionalProperties": false,
        "description": "How a server is authenticated \u2014 **the variable name, never the value**.\n\nThis model is the reason the panel can be honest. There is no field here\na credential could travel in, so the question \"did the editor just POST a\ntoken to the server\" has one answer and it is no: the browser never has\none, because the value lives in the server's own environment and is read\nat bind time by name.",
        "properties": {
          "headerName": {
            "default": "",
            "description": "`header` only \u2014 the vendor's own header name.",
            "title": "Headername",
            "type": "string"
          },
          "kind": {
            "default": "none",
            "description": "`none`, `bearer` or `header`.",
            "title": "Kind",
            "type": "string"
          },
          "tokenEnv": {
            "default": "",
            "description": "The NAME of the environment variable holding the credential.",
            "title": "Tokenenv",
            "type": "string"
          }
        },
        "title": "McpAuthPayload",
        "type": "object"
      },
...
      "McpServerResponse": {
        "description": "One registered server, as the editor may show it.",
        "properties": {
          "auth": {
            "$ref": "#/components/schemas/McpAuthPayload"
          },
          "credentialConfigured": {
            "description": "Whether the named variable is set on this server. Presence, not validity \u2014 only a handshake can tell you the second, which is what /validate is for.",
            "title": "Credentialconfigured",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "origin": {
            "description": "`built-in` for the two defaults, `project` for a config entry.",
            "title": "Origin",
            "type": "string"
          },
          "transport": {
            "title": "Transport",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "name",
          "url",
          "transport",
          "auth",
          "origin",
          "credentialConfigured"
        ],
        "title": "McpServerResponse",
        "type": "object"
      },
      "McpServerUsageResponse": {
        "description": "`GET /api/mcp/servers/{name}/usage` \u2014 who would break if it went away.",
        "properties": {
          "slugs": {
            "description": "Saved packages with a `tool.mcp` row naming this server.",
            "items": {
              "type": "string"
            },
            "title": "Slugs",
            "type": "array"
          }
        },
        "title": "McpServerUsageResponse",
        "type": "object"
      },
      "McpServerWriteRequest": {
        "additionalProperties": false,
        "description": "`POST /api/mcp/servers` \u2014 register one server in the project's config.\n\nThe same four facts `McpServerResponse` carries back, minus the two the\nserver owns: `origin` is decided by which file the entry lives in, and\n`credentialConfigured` is read from this process's environment. An editor\nthat could post either would be an editor that could claim a built-in was\na project entry.\n\n`auth` is the same `McpAuthPayload` the validate route takes, so the\n\"there is no field a credential fits in\" property is one schema and not\ntwo that could diverge.",
        "properties": {
          "auth": {
            "$ref": "#/components/schemas/McpAuthPayload",
            "default": {
              "headerName": "",
              "kind": "none",
              "tokenEnv": ""
            }
          },
          "name": {
            "description": "What a `tool.mcp` card names. Replaces an entry of the same name.",
            "title": "Name",
            "type": "string"
          },
          "transport": {
            "default": "streamable_http",
            "description": "`streamable_http` or `sse`.",
            "title": "Transport",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "name",
          "url"
        ],
        "title": "McpServerWriteRequest",
        "type": "object"
      },
      "McpValidateRequest": {
        "additionalProperties": false,
        "description": "`POST /api/mcp/validate` \u2014 check one server, configured or inline.\n\nEither name a registered server, or post a URL. Naming one that is not\nregistered is an error rather than a silent fall-through to the URL: the\ntwo mean different things and answering the wrong one is how a developer\nends up validating a server they are not about to bind.",
        "properties": {
          "auth": {
            "$ref": "#/components/schemas/McpAuthPayload",
            "default": {
              "headerName": "",
              "kind": "none",
              "tokenEnv": ""
            }
          },
          "server": {
            "default": "",
            "description": "A registered server by name.",
            "title": "Server",
            "type": "string"
          },
          "transport": {
            "default": "streamable_http",
            "description": "`streamable_http` or `sse`.",
            "title": "Transport",
            "type": "string"
          },
          "url": {
            "default": "",
            "description": "Inline: the server's URL.",
            "title": "Url",
            "type": "string"
          }
        },
        "title": "McpValidateRequest",
        "type": "object"
      },
      "McpValidateResponse": {
        "description": "The handshake's verdict, and the tool names it found.\n\n`tools` is what a developer actually wants from a validate button: the\ncount and the names are the proof the thing has a tool surface, which is\nwhat a `tool.mcp` node will bind. `initialize` alone would prove only\nthat something answered.",
        "properties": {
          "elapsedSeconds": {
            "default": 0.0,
            "title": "Elapsedseconds",
            "type": "number"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "serverName": {
            "default": "",
            "title": "Servername",
            "type": "string"
          },
          "serverVersion": {
            "default": "",
            "title": "Serverversion",
            "type": "string"
          },
          "status": {
            "description": "`live`, `unreachable`, `auth_required`, `not_mcp` or `not_installed` \u2014 never a stack trace. The last is the one verdict about this runtime rather than the server: the `[mcp]` extra is absent, so no socket was opened.",
            "title": "Status",
            "type": "string"
          },
          "tools": {
            "items": {
              "type": "string"
            },
            "title": "Tools",
            "type": "array"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "title": "McpValidateResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Everything the canvas editor and `/chat` do goes through the endpoints below \u2014 there is no private API, so a third client is a supported thing to build rather than a reverse-engineering exercise.\n\n**Two parts of the contract are not in this document, and cannot be.** `POST /api/runs/stream`, `POST /api/runs/resume` and `GET /api/events` are Server-Sent Event streams; OpenAPI has no vocabulary for a frame sequence, an event-name union, or the guarantee that exactly one of `done`/`interrupt`/`error` is the last frame. Those are written out in `docs/api.md`, which is also where the five calls a custom chat actually needs are shown end to end.\n\nThis document is generated from the app and committed at `docs/openapi.json`; the live one is served at `/openapi.json`.",
    "summary": "The HTTP surface both shipped UIs are built on.",
    "title": "OpenStateGraph runtime",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/mcp/servers": {
      "get": {
        "description": "The two built-in defaults, with the project's own layered over by name.\n\nBoth defaults are public and keyless \u2014 the LangChain documentation and the\nLangChain API reference \u2014 so a brand-new project has two working servers\nwithout configuring anything. A project entry sharing a name replaces the\ndefault it names and leaves the other alone.",
        "operationId": "mcp_servers_api_mcp_servers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/McpServerResponse"
                  },
                  "title": "Response Mcp Servers Api Mcp Servers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Which MCP servers this project can bind",
        "tags": [
          "Operations"
        ]
      },
      "post": {
        "description": "Write `mcp_servers:` in `openstategraph.yaml`, by name.\n\n**The comments around it survive**, which is not a nicety for this file:\nit is YAML precisely so the reasoning can sit beside the line it applies\nto, and a whole-document rewrite would delete all of it on the first Save\npress. `config_edit` replaces the one block and leaves every other byte.\n\n**The loader is the only judge of what may be written.** A `token_env`\nholding a pasted key is refused by re-reading the file through\n`load_config` \u2014 the same walk that refuses one on read \u2014 and the previous\nbytes go back. There is no second copy of the secret rules here to drift.\n\nAnswers with the whole list, not the entry: the panel's next question is\nalways \"what is registered now\", and a route that answers it saves a\nround trip and removes the state where the two disagree.",
        "operationId": "save_mcp_server_api_mcp_servers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpServerWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/McpServerResponse"
                  },
                  "title": "Response Save Mcp Server Api Mcp Servers Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register one MCP server in this project's config file",
        "tags": [
          "Operations"
        ]
      }
    },
...
    "/api/mcp/servers/{name}": {
      "delete": {
        "description": "Delete a project entry, or tombstone a built-in default.\n\nThe two defaults are not declared by any file, so removing one has no line\nto delete: it writes `enabled: false`, which the catalogue honours. That\nis the honest version of a Delete button that is offered on every row \u2014\nthe alternative, hiding it on the two rows a new project actually has,\nmakes the first two entries the only undeletable ones.",
        "operationId": "delete_mcp_server_api_mcp_servers__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/McpServerResponse"
                  },
                  "title": "Response Delete Mcp Server Api Mcp Servers  Name  Delete",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove one MCP server from this project",
        "tags": [
          "Operations"
        ]
      }
    },
    "/api/mcp/servers/{name}/restore": {
      "post": {
        "description": "Lift a tombstone. The other half of Delete on a `default` row.\n\nNot the same as re-registering it by hand: that writes a project entry, so\nthe server comes back wearing a `project` badge and a second copy of a URL\nthe product ships. This deletes the line, and the built-in is the built-in\nagain.",
        "operationId": "restore_mcp_default_api_mcp_servers__name__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/McpServerResponse"
                  },
                  "title": "Response Restore Mcp Default Api Mcp Servers  Name  Restore Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bring a hidden built-in default back",
        "tags": [
          "Operations"
        ]
      }
    },
    "/api/mcp/servers/{name}/usage": {
      "get": {
        "description": "Which saved documents would break if this server went away.\n\nA `tool.mcp` card **names** a server; the definition lives in the project's\nconfig. That indirection is the feature \u2014 a copied package carries no URL\nof yours \u2014 and it means deleting a server can silently break a document\nthat is not open, which nothing warned about (mcp-connect ticket 06).\n\nThe module docstring above says these routes do not touch a workflow, and\nthis one reads the package directory. The rule it was protecting is that a\nserver registry must not depend on a *workflow being open*, and this reads\nfiles rather than `WorkflowServices`, exactly as `_catalogue` reads the\nconfig file rather than a loaded config object. It is also asked only when\nsomebody is about to press Delete, never on a list.",
        "operationId": "mcp_server_usage_api_mcp_servers__name__usage_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerUsageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Saved workflows whose tool.mcp cards name this server",
        "tags": [
          "Operations"
        ]
      }
    },
    "/api/mcp/validate": {
      "post": {
        "description": "`initialize` + `list_tools`, wrapped in a 15-second ceiling.\n\n**It never refuses to answer.** Every outcome is a verdict with a status\nthe panel can render as a badge \u2014 that is the map's decision that\nvalidation *saves* with a live/unreachable/auth-required badge rather than\nblocking a developer who is mid-typing.\n\nThe ceiling is not caution. `DEFAULT_STREAMABLE_HTTP_SSE_READ_TIMEOUT` is\n**300 seconds**, so a server that accepts a connection and then says\nnothing would hold this request open for five minutes.\n\nMeasured against the public docs server: 0.86 s for both round trips.",
        "operationId": "validate_mcp_api_mcp_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpValidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Shake hands with one MCP server, and list what it offers",
        "tags": [
          "Operations"
        ]
      }
    },
...
  }
}